move getAttributes to firebase module, add mocks for tests
