+++ src/libs/Firebase/index.native.ts
@@ -2,12 +2,14 @@
 import crashlytics from '@react-native-firebase/crashlytics';
 import perf from '@react-native-firebase/perf';
 import * as Environment from '@libs/Environment/Environment';
+import * as PersonalDetailsUtils from '@libs/PersonalDetailsUtils';
+import * as ReportUtils from '@libs/ReportUtils';
 import * as SessionUtils from '@libs/SessionUtils';
-import type {Log, StartTrace, StopTrace, TraceMap} from './types';
+import type {FirebaseAttributes, Log, StartTrace, StopTrace, TraceMap} from './types';
 
 const traceMap: TraceMap = {};
 
-const startTrace: StartTrace = (customEventName, attributes) => {
+const startTrace: StartTrace = (customEventName) => {
     const start = global.performance.now();
     if (Environment.isDevelopment()) {
         return;
@@ -17,17 +19,14 @@ const startTrace: StartTrace = (customEventName, attributes) => {
         return;
     }
 
-    const session = SessionUtils.getSession();
+    const attributes = getAttributes();
 
     perf()
         .startTrace(customEventName)
         .then((trace) => {
-            trace.putAttribute('accountID', session?.accountID?.toString() ?? 'N/A');
-            if (attributes) {
-                Object.entries(attributes).forEach(([name, value]) => {
-                    trace.putAttribute(name, value);
-                });
-            }
+            Object.entries(attributes).forEach(([name, value]) => {
+                trace.putAttribute(name, value);
+            });
             traceMap[customEventName] = {
                 trace,
                 start,
@@ -59,6 +58,20 @@ const stopTrace: StopTrace = (customEventName) => {
 const log: Log = (action: string) => {
     crashlytics().log(action);
 };
+
+function getAttributes(): FirebaseAttributes {
+    const session = SessionUtils.getSession();
+
+    const accountId = session?.accountID?.toString() ?? 'N/A';
+    const reportsLength = ReportUtils.getAllReportsLength().toString();
+    const personalDetailsLength = PersonalDetailsUtils.getPersonalDetailsLength().toString();
+
+    return {
+        accountId,
+        reportsLength,
+        personalDetailsLength,
+    };
+}
 
 export default {
     startTrace,
