+++ src/libs/Navigation/AppNavigator/AuthScreens.tsx
@@ -11,7 +11,6 @@ import useStyleUtils from '@hooks/useStyleUtils';
 import useThemeStyles from '@hooks/useThemeStyles';
 import useWindowDimensions from '@hooks/useWindowDimensions';
 import {READ_COMMANDS} from '@libs/API/types';
-import getAttributes from '@libs/Firebase/utils';
 import HttpUtils from '@libs/HttpUtils';
 import KeyboardShortcut from '@libs/KeyboardShortcut';
 import Log from '@libs/Log';
@@ -211,7 +210,7 @@ function AuthScreens() {
     let initialReportID: string | undefined;
     const isInitialRender = useRef(true);
     if (isInitialRender.current) {
-        Timing.start(CONST.TIMING.HOMEPAGE_INITIAL_RENDER, getAttributes());
+        Timing.start(CONST.TIMING.HOMEPAGE_INITIAL_RENDER);
 
         const currentURL = getCurrentUrl();
         if (currentURL) {
