+++ src/libs/Navigation/AppNavigator/createCustomBottomTabNavigator/TopBar.tsx
@@ -10,7 +10,6 @@ import WorkspaceSwitcherButton from '@components/WorkspaceSwitcherButton';
 import useLocalize from '@hooks/useLocalize';
 import useTheme from '@hooks/useTheme';
 import useThemeStyles from '@hooks/useThemeStyles';
-import getAttributes from '@libs/Firebase/utils';
 import Navigation from '@libs/Navigation/Navigation';
 import Performance from '@libs/Performance';
 import SignInButton from '@pages/home/sidebar/SignInButton';
@@ -66,7 +65,7 @@ function TopBar({breadcrumbLabel, activeWorkspaceID, shouldDisplaySearch = true}
                             accessibilityLabel={translate('sidebarScreen.buttonFind')}
                             style={[styles.flexRow, styles.mr2, styles.touchableButtonImage]}
                             onPress={Session.checkIfActionIsAllowed(() => {
-                                Timing.start(CONST.TIMING.CHAT_FINDER_RENDER, getAttributes());
+                                Timing.start(CONST.TIMING.CHAT_FINDER_RENDER);
                                 Performance.markStart(CONST.TIMING.CHAT_FINDER_RENDER);
                                 Navigation.navigate(ROUTES.CHAT_FINDER);
                             })}
