+++ src/libs/OptionsListUtils.ts
@@ -42,7 +42,6 @@ import times from '@src/utils/times';
 import Timing from './actions/Timing';
 import * as ErrorUtils from './ErrorUtils';
 import filterArrayByMatch from './filterArrayByMatch';
-import getAttributes from './Firebase/utils';
 import localeCompare from './LocaleCompare';
 import * as LocalePhoneNumber from './LocalePhoneNumber';
 import * as Localize from './Localize';
@@ -2078,7 +2077,7 @@ function getOptions(
  * Build the options for the Search view
  */
 function getSearchOptions(options: OptionList, searchValue = '', betas: Beta[] = []): Options {
-    Timing.start(CONST.TIMING.LOAD_SEARCH_OPTIONS, getAttributes());
+    Timing.start(CONST.TIMING.LOAD_SEARCH_OPTIONS);
     Performance.markStart(CONST.TIMING.LOAD_SEARCH_OPTIONS);
     const optionList = getOptions(options, {
         betas,
