+++ src/pages/home/report/ReportActionItemParentAction.tsx
@@ -5,7 +5,6 @@ import OfflineWithFeedback from '@components/OfflineWithFeedback';
 import TripDetailsView from '@components/ReportActionItem/TripDetailsView';
 import useNetwork from '@hooks/useNetwork';
 import useThemeStyles from '@hooks/useThemeStyles';
-import getAttributes from '@libs/Firebase/utils';
 import Navigation from '@libs/Navigation/Navigation';
 import onyxSubscribe from '@libs/onyxSubscribe';
 import * as ReportActionsUtils from '@libs/ReportActionsUtils';
@@ -136,7 +135,7 @@ function ReportActionItemParentAction({
                                               // Pop the chat report screen before navigating to the linked report action.
                                               Navigation.goBack(ROUTES.REPORT_WITH_ID.getRoute(ancestor.report.parentReportID ?? '-1', ancestor.reportAction.reportActionID));
                                           }
-                                          Timing.start(CONST.TIMING.SWITCH_REPORT, getAttributes());
+                                          Timing.start(CONST.TIMING.SWITCH_REPORT);
                                       }
                                     : undefined
                             }
